+++ Locales/ptBR.lua
@@ -238,6 +238,8 @@ L["Arrow"] = "Seta"
 L["Arrow (Legacy)"] = "Seta (Legado)"
 L["Art Options"] = "Opções de Arte"
 --[[Translation missing --]]
+L["Artwork"] = "Artwork"
+--[[Translation missing --]]
 L["Attempt to register script for unknown WoW event \"%s\""] = "Attempt to register script for unknown WoW event \"%s\""
 --[[Translation missing --]]
 L["Attempt to register script for unknown WoW event '%s'"] = "Attempt to register script for unknown WoW event '%s'"
@@ -247,6 +249,8 @@ L["Aura: "] = "Aura:"
 L["Auras"] = "Auras"
 L["Auras (Name or ID)"] = "Auras (Nome ou ID)"
 L["Auras, Healthbar"] = "Auras, Barra de Vida"
+--[[Translation missing --]]
+L["Authors"] = "Authors"
 L["Auto Sizing"] = "Dimensionamento automático"
 L["Auto-Cast"] = "Lançamento automático"
 --[[Translation missing --]]
@@ -368,6 +372,10 @@ L["Class Color"] = "Class Color"
 L["Class Color for Players"] = "Class Color for Players"
 L["Class Icon"] = "Ícone de classe"
 --[[Translation missing --]]
+L["Clean"] = "Clean"
+--[[Translation missing --]]
+L["Clean Border"] = "Clean Border"
+--[[Translation missing --]]
 L[ [=[Clear and easy to use threat-reactive nameplates.
 
 Current version: ]=] ] = [=[Clear and easy to use threat-reactive nameplates.
@@ -410,12 +418,16 @@ L["Configuration Mode"] = "Configuration Mode"
 --[[Translation missing --]]
 L["Controls the rate at which nameplate animates into their target locations [0.0-1.0]."] = "Controls the rate at which nameplate animates into their target locations [0.0-1.0]."
 --[[Translation missing --]]
+L["Cooldown"] = "Cooldown"
+--[[Translation missing --]]
 L["Cooldown Spiral"] = "Cooldown Spiral"
 L["Copy"] = "Copiar"
 --[[Translation missing --]]
 L["Creation"] = "Creation"
 --[[Translation missing --]]
 L["Crescent"] = "Crescent"
+--[[Translation missing --]]
+L["Crest"] = "Crest"
 --[[Translation missing --]]
 L["Crowd Control"] = "Crowd Control"
 --[[Translation missing --]]
@@ -641,6 +653,10 @@ L["Event Name"] = "Event Name"
 L["Events with Script"] = "Events with Script"
 L["Everything"] = "Tudo"
 --[[Translation missing --]]
+L["Evoker"] = "Evoker"
+--[[Translation missing --]]
+L["Evoker Essence Cooldown"] = "Evoker Essence Cooldown"
+--[[Translation missing --]]
 L["Exchange"] = "Exchange"
 --[[Translation missing --]]
 L["Experience"] = "Experience"
@@ -666,6 +682,8 @@ L["Fading"] = "Fading"
 L["Failed to migrate the imported profile to the current settings format because of an internal error. Please report this issue at the Threat Plates homepage at CurseForge: "] = "Failed to migrate the imported profile to the current settings format because of an internal error. Please report this issue at the Threat Plates homepage at CurseForge: "
 --[[Translation missing --]]
 L["Fanged Pit Viper"] = "Fanged Pit Viper"
+--[[Translation missing --]]
+L["FAQ"] = "FAQ"
 --[[Translation missing --]]
 L["Filter"] = "Filter"
 --[[Translation missing --]]
@@ -702,6 +720,8 @@ L["Format"] = "Format"
 L["Four"] = "Four"
 --[[Translation missing --]]
 L["Frame Order"] = "Frame Order"
+--[[Translation missing --]]
+L["Frequently Asked Questions"] = "Frequently Asked Questions"
 --[[Translation missing --]]
 L["Friendly & Neutral Units"] = "Friendly & Neutral Units"
 --[[Translation missing --]]
@@ -1244,6 +1264,8 @@ L["Sets your role to tanking."] = "Sets your role to tanking."
 --[[Translation missing --]]
 L["Sets your spec "] = "Sets your spec "
 --[[Translation missing --]]
+L["Seven"] = "Seven"
+--[[Translation missing --]]
 L["Shadow"] = "Shadow"
 --[[Translation missing --]]
 L["Shadow Fiend"] = "Shadow Fiend"
@@ -1683,6 +1705,8 @@ L["Threat Detection Heuristic"] = "Threat Detection Heuristic"
 --[[Translation missing --]]
 L["Threat Glow"] = "Threat Glow"
 --[[Translation missing --]]
+L["Threat Plates is currently maintained by Backup (CurseForge). It was originally written by Suicidal Katt (CurseForge)."] = "Threat Plates is currently maintained by Backup (CurseForge). It was originally written by Suicidal Katt (CurseForge)."
+--[[Translation missing --]]
 L["Threat Plates Script Editor"] = "Threat Plates Script Editor"
 --[[Translation missing --]]
 L["Threat System"] = "Threat System"
@@ -1729,10 +1753,14 @@ L["Totem Transparency"] = "Totem Transparency"
 --[[Translation missing --]]
 L["Totems"] = "Totems"
 --[[Translation missing --]]
+L["Translators"] = "Translators"
+--[[Translation missing --]]
 L["Transliterate Cyrillic Letters"] = "Transliterate Cyrillic Letters"
 L["Transparency"] = "Transparência"
 --[[Translation missing --]]
 L["Transparency & Scaling"] = "Transparency & Scaling"
+--[[Translation missing --]]
+L["Transparent"] = "Transparent"
 --[[Translation missing --]]
 L["Treant"] = "Treant"
 --[[Translation missing --]]
@@ -1851,6 +1879,8 @@ L["Value Type"] = "Value Type"
 --[[Translation missing --]]
 L["Venomous Snake"] = "Venomous Snake"
 --[[Translation missing --]]
+L["Version"] = "Version"
+--[[Translation missing --]]
 L["Vertical Align"] = "Vertical Align"
 --[[Translation missing --]]
 L["Vertical Alignment"] = "Vertical Alignment"
@@ -1887,6 +1917,10 @@ L["With Pet"] = "With Pet"
 L["Word Wrap"] = "Word Wrap"
 --[[Translation missing --]]
 L["World Boss"] = "World Boss"
+--[[Translation missing --]]
+L["WoW Flat"] = "WoW Flat"
+--[[Translation missing --]]
+L["WoW Round"] = "WoW Round"
 L["X"] = "X"
 L["Y"] = "Y"
 --[[Translation missing --]]